+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2]
+
+### Added
+
+- Mongo handler
+
 ## [0.0.1] - `TODO` Add release date
 
 ### Added
 
-- `application-environment-express-http`
+- `application-environment-express-http` - Consumes
   - Setup express server for local development. [Tracking](https://github.com/creekburn/server/issues/1)
   - `TODO` Map OpenAPI specification to routing information. [Tracking](https://github.com/creekburn/server/issues/4)
   - `TODO` Map routing information to express api. [Tracking](https://github.com/creekburn/server/issues/3)